+++ controller/option.go
@@ -159,7 +159,7 @@ func UpdateOption(c *gin.Context) {
 			})
 			return
 		}
-	case "QuotaForInviter", "QuotaForInvitee":
+	case "QuotaForInviter", "QuotaForInvitee", "ReferralDepositPercent":
 		if isPositiveOptionValue(option.Value.(string)) && !operation_setting.IsPaymentComplianceConfirmed() {
 			common.ApiErrorI18n(c, i18n.MsgPaymentComplianceRequired)
 			return